% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/brca_reference_regressions.R
\docType{data}
\name{brca_reference_regressions}
\alias{brca_reference_regressions}
\title{BRCA reference regressions}
\format{
List containing cpg variance, and regression slopes, intercepts, resiudal standard error and degrees of freedom
}
\source{
<The Cancer Genome Atlas>
}
\usage{
brca_reference_regressions
}
\description{
List containing the parameters of the reference regressions obtained from the TCGA-BRCA dataset.
}
\keyword{datasets}
