% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/example_betas_to_correct.R
\docType{data}
\name{example_betas_to_correct}
\alias{example_betas_to_correct}
\title{example_betas_to_correct}
\format{
A matrix with 5000 rows (CpGs) and 504 columns (samples).
}
\source{
<https://github.com/StaafLab/PureBeta>
}
\usage{
example_betas_to_correct
}
\description{
A matrix containing the uncorrected beta values of the 5.000 most variable CpGs from the TCGA-BRCA dataset. It is intended to be used as an example file to run the beta correction based on the reference files.
}
\keyword{datasets}
