% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/example_purities_reference.R
\docType{data}
\name{example_purities_reference}
\alias{example_purities_reference}
\title{example_purities_reference}
\format{
A named vector with 126 elements.
}
\source{
<https://github.com/StaafLab/PureBeta>
}
\usage{
example_purities_reference
}
\description{
A named vector with the purity values of the samples included in the reference beta value matrix. It is intended to be used as a reference to correct betas or estimate purity based on reference regressions.
}
\keyword{datasets}
