% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/luad_reference_regressions.R
\docType{data}
\name{luad_reference_regressions}
\alias{luad_reference_regressions}
\title{LUAD reference regressions}
\format{
List containing cpg variance, and regression slopes, intercepts, resiudal standard error and degrees of freedom
}
\source{
<The Cancer Genome Atlas>
}
\usage{
luad_reference_regressions
}
\description{
List containing the parameters of the reference regressions obtained from the TCGA-LUAD dataset.
}
\keyword{datasets}
