% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lusc_reference_regressions.R
\docType{data}
\name{lusc_reference_regressions}
\alias{lusc_reference_regressions}
\title{LUSC reference regressions}
\format{
List containing cpg variance, and regression slopes, intercepts, resiudal standard error and degrees of freedom
}
\source{
<The Cancer Genome Atlas>
}
\usage{
lusc_reference_regressions
}
\description{
List containing the parameters of the reference regressions obtained from the TCGA-LUSC dataset.
}
\keyword{datasets}
