% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/beta_correction_for_cohorts.R
\name{beta_correction_for_cohorts}
\alias{beta_correction_for_cohorts}
\title{beta_correction_for_cohorts}
\usage{
beta_correction_for_cohorts(
  beta_values,
  tumour_purities,
  set_seed = FALSE,
  seed_num = 2000,
  cores = 1
)
}
\arguments{
\item{beta_values}{A matrix with CpGs as rows and tumour samples as columns
with the uncorrected beta values from the CpGs of the samples that are
intended to be corrected. The values must be numeric, the rows must be named
with the CpG ID, and the columns with the sample IDs. An example of the
required format is available in the example_betas_reference matrix.}

\item{tumour_purities}{Named vector containing the sample purity values of
of the samples whose DNA methylation beta values are intended to be corrected.
The vector must be named with the sample ID, which must match with the sample
IDs from the matrix containing the beta values. An example of the required
format is available in the example_purities_reference vector.}

\item{set_seed}{Default = FALSE. A seed for the FlexMix package to detect
the different CpG methylation patterns can be used by setting this argument
to TRUE.}

\item{seed_num}{Default = 2000. The seed to be used when set_seed = TRUE can
be specified here.}

\item{cores}{Default = 1. Number of cores to be used to run the function in
parallel.}
}
\value{
List containing the original uncorrected beta values
(output$betas.original), the corrected tumour beta values (output$betas.tumour)
and the corrected microenvironment beta values (output$betas.microenvironment).
}
\description{
This functions applies the original Staaf & Aine DNA methylation beta value
correction strategy to adjust the effect of sample purities in cohorts with
known sample purities and CpG beta values (Staaf & Aine, Plos One, 2022).
This function allows mult-core execution.
}
\examples{

# Using the default parameters
beta_correction_for_cohorts (beta_values = example_betas_reference,
                             tumour_purities = example_purities_reference)

# Specifying new parameters
beta_correction_for_cohorts (beta_values = example_betas_reference,
                             tumour_purities = example_purities_reference,
                             set_seed = TRUE,
                             seed_num = 1,
                             cores = 5)
}
