% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/example_betas_reference.R
\docType{data}
\name{example_betas_reference}
\alias{example_betas_reference}
\title{example_betas_reference}
\format{
A matrix with 5000 rows (CpGs) and 126 columns (samples).
}
\source{
<https://github.com/StaafLab/PureBeta>
}
\usage{
example_betas_reference
}
\description{
A matrix containing the uncorrected beta values of the 5.000 most variable CpGs from the TCGA-BRCA dataset. This matrix is intended to be used as example reference data for the PureBeta package.
}
\keyword{datasets}
